Refactor: 셀프클로징, import, props 등 코딩 컨벤션 적용
